chore: Move doctypes declaration file
