+++ targets/drive/web/services.jsx
@@ -10,7 +10,7 @@ import CozyClient, { CozyProvider } from 'cozy-client'
 import { I18n } from 'cozy-ui/react/I18n'
 import { getQueryParameter } from 'react-cozy-helpers'
 import { upgradePouchDatabase } from 'drive/lib/upgradePouchDatabase'
-import { schema } from '../doctypes'
+import { schema } from 'drive/lib/doctypes'
 if (__DEVELOPMENT__) {
   // Enables React dev tools for Preact
   // Cannot use import as we are in a condition
